Replace ConvPartial vertex with ConvPartialHorizontalMac vertex.
Summary:
The logic that created ConvPartial vertices was hardcoded to add
one vertex per output group and never split the x-axis. The way the
ConvPartial vertex represented tasks made it hard to lift this restriction.
Replace ConvPartial with a new vertex type (ConvPartialHorizontalMac) that
has a separate weight, input and output pointers for each pass over the
partial sums.

Using separate pointers for each pass over the partial sums moves logic
from the vertex to the host code that creates the graph. This provides
more flexibility and makes it easier to split the worker between vertices
in different ways. A disadvantage of this approach is that the number
of edge pointers increases.

Currently the ConvPartial vertex is not used in ResNet / Alexnet so this
change does not affect the benchmarks. The motivation for this change is
that when you recast the fully connected layer as a convolution then you
end up with a convolution where the number of output channels is equal to
the batch size of the fully connected layer. For small numbers of output
channels (e.g. 1) the refusal to split the x-axis made the old code
was very inefficient.

Reviewers: davel

Reviewed By: davel

Differential Revision: https://phabricator.sourcevertex.net/D1552