Update fully_connected_layer to use popconv instead of poplin.
Summary:
The forward pass is implemented as a convolution where the activations are
loaded into the AMP unit. The number of output channels in the convolution
is equal to the batch size of the fully connected layer. The backward pass
is implemented as a weight update with a plan derived from the forward
pass to avoid rearrangement of weights.

The fully connected weight update still uses matMul. As noted in the
TODO comment this is inefficient.

Reviewers: davel

Reviewed By: davel

Differential Revision: https://phabricator.sourcevertex.net/D1575