Remove verticesPerTilePerYAxis from the plan.
Summary:
This member is no longer used to decide how many vertices to create.
In order to remove verticesPerTilePerYAxis from the plan I've assumed a
verticesPerTilePerYAxis of 1 inside the estimatePartialCalcMemory() function.
This is incorrect but the function was incorrect before this change as it is
no longer in sync with implementation of the convolution function.

Reviewers: davel

Reviewed By: davel

Differential Revision: https://phabricator.sourcevertex.net/D1591