Change comparison used when the memory / cycle bounds are set.
Summary:
Previously if primaryCheckIsCycles was true the memory bounds were
ignored and similarly if it was false the cycles bound were ignored.
After this changes the comparison takes both bounds are taken into account.
If primaryCheckIsCycles is true we compare solutions as follows:

1. If one solution has an in bounds cycle count and the other doesn't
   prefer the solution that is in bounds.
2. Otherwise if one solution has an in bounds memory and the other doesn't
   prefer the solution that is in bounds.
3. Otherwise prefer the solution that uses fewer cycles.
4. Otherwise prefer the solution that uses less memory.

Similar logic applies if primaryCheckIsCycles is set to false with memory
cycles swapped around.

When the --percent-cyc-excess-for-mem-optim option is specified we want to
minimize the memory usage (subject to the cycle constraint) and with the
new comparison logic we must set primaryCheckIsCycles to false to do this.

Reviewers: godfrey.da.costa

Reviewed By: godfrey.da.costa

Differential Revision: https://phabricator.sourcevertex.net/D1638