Separate bias addition from the conv layer
Summary:
For the forward pass, the bias addition is done by creating a tensor which is
repeated for each field position. This allows using popstd::addTo library
function but has a potential memory overhead. If needed the operation can be
specialised.

The bias update function in weight-update pass is provided as a popconv library
function.

Fixes Maniphest Task T1027

Test Plan: Existing tests are affected  by this change

Reviewers: davel

Reviewed By: davel

Differential Revision: https://phabricator.sourcevertex.net/D1633