Specify the input and output stride as fields.
Summary:
Previously input and output strides were inferred from the sizes of the
inputs and outputs on the assumption that at most one stride was not 1.

The change opens the door to striding the input and output at the same time
(although the host code doesn't target this yet). My intent is to use this
to arrange outputs in groups of 16. We can avoid a rearrangement step after
the convolution if we use an output stride of 2. If the convolutional
layer uses input striding as well we must support striding both the input
and the output in the vertex.

Depends on D1644

Reviewers: godfrey.da.costa

Reviewed By: godfrey.da.costa

Differential Revision: https://phabricator.sourcevertex.net/D1646