Update convolution planning to use popsolver.
Summary:
choosePlan() uses popsolver to model the problem of finding a plan as a
constraint satisfaction problem. If bounds on cycles or memory are
specified they are added as separate constraints.

There is a small change of functionality. The old code didn't loop over
every possible value of tilesPerInZ but it instead made tilesPerInZ as
large as possible given the values of the other variables. The new code
now explores the full search space.

This change increases the compile time (mostly due to the larger search
space). The time to run benchmarks/resnet-18 increases from 23.50s to
32.55s

Fixes T1181
Depends on D1631, D1638

Reviewers: davel

Reviewed By: davel

Maniphest Tasks: T1181

Differential Revision: https://phabricator.sourcevertex.net/D1640