Avoid an unnecessary cast at the end of a convolution.
Summary:
convolutionByAmp() produces an output with the same type as the input type
so there is no need to perform a cast afterwards.

If plan.flattenXY is true we return dimshuffled view of the activations
tensor - this seems preferable to inserting a copy. This depends on D1661
which adds support for device <-> host copies of non contiguous tensors.

Reviewers: davel

Reviewed By: davel

Differential Revision: https://phabricator.sourcevertex.net/D1662