Generalize backward pass for residuals
Summary:
Previously the backwards pass through a residual add did not create
gradients with shapes that matched the inputs. This was then dealt with
by special logic in the code that used those gradients. This change
creates the correct input tensor gradients, removes the special logic
and replaces it with a general summation for outputs of operations that
are used by multiple operations.

Reviewers: Eddy

Reviewed By: Eddy

Differential Revision: https://phabricator.sourcevertex.net/D1690