Support larger partial channel groupings.
Summary:
It was previously assumed that the partial channel grouping was equal
to the number of convolutional units. With this change the partial channel
grouping is allowed to be any multiple of the number of convolutional
units. Output striding in the vertex is used if the partial channel
grouping is larger than the number of convolutional units.

This logic is not yet exercised as the the planner still chooses the
channel grouping based on the number of convolutional units.

Reviewers: davel

Reviewed By: davel

Differential Revision: https://phabricator.sourcevertex.net/D1664