Fix segfault caused by dereference to nullptr when accessing ConvOptions::cache
Summary:
Fixes T1224. Fix segfault caused by attempting to dereference  ConvOptions::cache when it is a nullptr.

This occurs in the case where a default constructed ConvOptions is provided to the getPlan, getWeightUpdatePlan method, which has a pointer to PlanningCache, which the methods attempt to dereference.

The getPlan and getWeightUpdatePlan functions will now create a temporary PlanningCache at function scope, before trying to create the plan.

Reviewers: davel, davidn, richard

Reviewed By: richard

Maniphest Tasks: T1224

Differential Revision: https://phabricator.sourcevertex.net/D1740