Add unary and binary operators in poplibs
Summary:
Only variants of operations which create and return an output tensor are
defined. A further rationalisation of variants of the operators must be done.
The compute cost for pow(a,b) is not correct.

Cycle cost for operators assume that dual load with store cannot be employed
as the load/store of may refer to the same tensor.

Fixes Maniphest T1182

Test Plan: New test added with subtest for each operator

Reviewers: Eddy

Reviewed By: Eddy

Subscribers: davidn

Differential Revision: https://phabricator.sourcevertex.net/D1745