Initialise inputs to conv layer tool to have a positive mean
Summary:
With inputs initialised with samples from an uniform distribution with zero
mean and large standard deviation, convolutions with a large number of
constituent reductions have a higher chance of cancellation resulting in values
close to zero. One solution is to bias the mean to be positive to reduce the
probablility of cancellation. Other approaches are discussed in Maniphest task
T1277

Test Plan: Tested kernels ox size 3x3 and input channels of size 4096

Reviewers: Eddy

Reviewed By: Eddy

Differential Revision: https://phabricator.sourcevertex.net/D1826