Make use of iterateTilePartition() in calculateWeightDeltasAop().
Summary:
The previous attempt to use iterateTilePartition() (commit dd905668) had
to be reverted as the weight update pass didn't transform the ConvParams
structure in the same was as the forward pass did. This change updates
calculateWeightDeltasAop() so it transforms the plan and the parameters
in the same way as the forward pass and switches over to
iterateTilePartition() in order to iterate over the partition.

Reviewers: davel

Reviewed By: davel

Differential Revision: https://phabricator.sourcevertex.net/D1968