Support using AMP instead of AOP in the fully connected backward pass.
Summary:
The fullyConnectedBwdAsWU option is used to control whether AMP or AOP is
used. If it is set to false the fully connected backward pass is cast as
a convolution. Between the forward and backward pass the weights must be
transposed. We choose a joint plan that avoids the need to exchange
weights between tiles. The transposition of the weights uses the existing
Transpose2D vertex type.

Fixes T1290
Depends on D1997

Reviewers: davel

Reviewed By: davel

Maniphest Tasks: T1290

Differential Revision: https://phabricator.sourcevertex.net/D1998