Fix weightUpdateByAmpTransformParams() when striding is used.
Summary:
If there are excess elements at the end that are ignored when the output
is subsampled in the forward pass then this should be taken into account
when backprogating errors during the weight update pass. Introduce a helper
function that transforms the convolution into an equivalent convolution
with no ignored elements at the end to simplify the code.

Depends on D2039, D2046, D2051

Reviewers: davel

Reviewed By: davel

Differential Revision: https://phabricator.sourcevertex.net/D2052