Optimisations to reduce cycle and memory usage for batch normalisation
Summary:
Retain inverse of standard deviation rather than standard deviation itself. The reason for this is that  the eps used could be modified to ensure that the inverse always fits within half. It should be noted that retaining standard deviation gives  a lot more dynamic range because half float denorms can be used to extend dynamic range. In the unlikely case that eps  < 1e-9 is used, we could use float precision storage for inverse standard deviation.

Modify the reduction function to use tile introspection especially because assumptions on activation mapping of output of convolutions cannot be made.

Some of the functions which can be in a single compute set now belong to a single compute set.

Reviewers: davel

Reviewed By: davel

Differential Revision: https://phabricator.sourcevertex.net/D2167