Add a function in popreduce to reduce first dimension and scale
Summary:
A reduction operation followed by a scale is added in popreduce. This is useful for instance in computing the mean of a tensor along a dimension.

Fixes Maniphest T1452

Test Plan:
Created a new test called ReductionTests tjat includes subtests for existing
reduction functions and tests for the new function.

Reviewers: Eddy

Reviewed By: Eddy

Differential Revision: https://phabricator.sourcevertex.net/D2239