Set variables that are the operands of call constraints first.
Summary:
In cases where the problem is thinly spread then, once an initial
solution had been found, the domain of the variable denoting the cost
might be smaller than the domain of the variables that control the split
across tiles. Previously the constraint solver would set the cost variable
before setting the other variables. Since setting the cost does not
allow it to reduce the domain of the other variables this just increased
execution time.

Fixes T1578

Reviewers: davel

Reviewed By: davel

Maniphest Tasks: T1578

Differential Revision: https://phabricator.sourcevertex.net/D2290