Optimize convolutions which are outer products.
Summary:
fully_connected_layer recasts the weight update of a fully connected
layer as a 1x1 convolution with b input channels where b is the batch size.
When the batch size is 1 this is equivalent to an outer product. The
existing methods of calculating convolutions where inefficient in this
case. Add a new method (OUTER_PRODUCT) to address this.

Fixes T1296

Reviewers: godfrey.da.costa

Reviewed By: godfrey.da.costa

Maniphest Tasks: T1296

Differential Revision: https://phabricator.sourcevertex.net/D2303