Generalise popreduce to support other operations
Summary:
A new library function is added to support operations ADD, MUL, MIN, MAX, OR
and AND. The dimensions to reduce is an argument to the function and can be
empty or equal to the rank of the input tensor to reduce.

Given the generic nature of the function, an optimal graph may require a
multi-stage reduce. The current implementation uses a single stage reduce.

Fixes Maniphest T1047

Test Plan: Added new tests in ReductionTests

Reviewers: davel

Reviewed By: davel

Subscribers: davidn

Differential Revision: https://phabricator.sourcevertex.net/D2283