Update reduce cycle estimate to reflect the implementation.
Summary:
The reduction of partial sums for an output volume is spread over the
tiles that store partial sums for that output volume. Update the
cost estimate to reflect this.

Although this change does affect the estimated cycle count it currently
has no effect on the partition chosen for ResNet-18 or ResNet-50.

Reviewers: godfrey.da.costa

Reviewed By: godfrey.da.costa

Differential Revision: https://phabricator.sourcevertex.net/D2331