Treat the batch axis more like a spatial axis.
Summary:
Previously the handling of the batch axis depended on the size of the
kernel. For n x n kernels (n > 1) the batch axis was always split across
tiles to the maximum extent possible - we never processed two different
elements of the batch on the same tile. For 1x1 kernels we partially
flattened the batch axis with the field axes depending on the value of
batchesPerGroup.

This change treats the batch axis more like the x / y axis. The
batchesPerGroup member of the plan has been replaced the tilesPerBatchAxis
member that is analogous to tilesPerXAxis or tilesPerYAxis. Previously
the calcPartialConvOutput() function operated on a single batch and was
passed slices of the input and output tensors where the batch dimension
has been stripped. Now it is passed input and output tensors with a batch
dimension and given a range of batch elements to operate on. For 1x1
convolutions the batch axis is always flattened into field since it gives
the planner the most freedom and there is no downside.

These changes mean the planner can now pick plans that wouldn't have been
possible previously. The number of cycles for ResNet-50 batch size 4 is
reduced from 1290795 to 1173605.

Ref T1435

Reviewers: godfrey.da.costa

Reviewed By: godfrey.da.costa

Maniphest Tasks: T1435

Differential Revision: https://phabricator.sourcevertex.net/D2334