Fix out of bound container access bug in reduce function
Summary:
In certain cases the vertex mapping vector could be smaller than the number of tiles in the device. To ensure out of bound containers are not accessed, the loop which divides vertices on tiles must do so only for tiles for which mapping info is available.

Fixes Maniphest T1660

Reviewers: richard

Reviewed By: richard

Differential Revision: https://phabricator.sourcevertex.net/D2362