Set the type of field as expected by random generator vertex
Summary:
There is a bug in poplar because of which value passed to set initial value of field may not do the correct conversions. This is probably due to the way rvalues and lvalues are processed.

Explicitly set the value expected by the vertex in a lvalue before call to setInitialValue. Also ensure that the full range is supported in integer uniform generator.

The specification of uniform number generator now requires a closed interval for all types.

Test Plan: Added new test which covers max integer range for uniform random number generator

Reviewers: Eddy

Reviewed By: Eddy

Differential Revision: https://phabricator.sourcevertex.net/D2364