Specify weights as the right hand side of the matrix multiplication.
Summary:
Because the batch dimension is the outer dimension of the activations
tensor it is more natural to describe the fully connected layer as a
matrix multiplication where the activations are the left operand and the
weights are the right operand. Previously we called matMul with the
operands the other way round which meant we had to transpose the
activations before calling matMul() and to tranpose the result of matMul().

Fixes T1631

Reviewers: godfrey.da.costa

Reviewed By: godfrey.da.costa

Maniphest Tasks: T1631

Differential Revision: https://phabricator.sourcevertex.net/D2387