Fix mac cycle estimate.
Summary:
When calculating the size of the convolution round up instead of down.
Previously the planner would exploit the fact that we rounded down by
choosing plans where the dividend is just less that the divisor as this
resulted in an size of zero and therefore an estimated cost of zero.

Reviewers: davel

Reviewed By: davel

Differential Revision: https://phabricator.sourcevertex.net/D2417