Support expanding spatial fields to increase the number input channels.
Summary:
We can increase the number of input channels by treating activations
multiplied different positions in the kernel as different input channels.
This transformation is necessary to efficiently implement a
convolutional weight update that is represented as a convolution.

For ResNet-50 inference the planner decides to expand dimensions for 4
of layers in ResNet-50 (mainly those with a stride of 2) and this results
in an overall speed up from 465219 cycles -> 454825 cycles

Ref T1508

Reviewers: godfrey.da.costa

Reviewed By: godfrey.da.costa

Maniphest Tasks: T1508

Differential Revision: https://phabricator.sourcevertex.net/D2447