Generalize the flattening of dimensions.
Summary:
Previously we either flattened all the dimensions or none of them.
Extend the plan to allow flattening of a subset of the dimensions. This is
useful in conjunction with expansion of dimensions - if a subset of the
dimensions are expanded we can always flatten expanded dimensions but we
may not be able flatten the other dimensions.

Depends on D2447

Reviewers: godfrey.da.costa

Reviewed By: godfrey.da.costa

Differential Revision: https://phabricator.sourcevertex.net/D2452