Support flattening spatial dimensions into the output channels.
Summary:
When the kernel size is bigger than the field size then when we expand the
spatial dimension it results in a field size of 1, i.e. the same
activations are used for each position in the output. This is equivalent to
having extra output channels and we can flatten the dimension into the
output channels.

Depends on D2452

Reviewers: godfrey.da.costa

Reviewed By: godfrey.da.costa

Differential Revision: https://phabricator.sourcevertex.net/D2460