Generalize getExpandDimsCandidates() to an arbitrary number of dimensions.
Summary:
Extend getExpandDimsCandidates() to work for an arbitrary number of
spatial dimensions. Only consider expanding a dimension if it would
change the parameters. This improves compile time by reducing the
number of candidates we must evaluate. It also simplifies
estimateTransformCost() as it no longer needs to check for a
transformations that have no cost due to being a no-op. This fixes a bug
where estimateTransformCost() considered an expansion made the input
dilation / padding explict to have no cost if it didn't change the number
of input channels.

Reviewers: davel

Reviewed By: davel

Differential Revision: https://phabricator.sourcevertex.net/D2485