Improve choice of which tensor to expand.
Summary:
When a spatial dimension is expanded we either expand the activations or
the weights, whichever is larger (after dilation and padding). If the
activations and the weights are the same size after dilation and padding
we have a free choice over which one is expanded. We are better off
expanding the that has the least amount of zero padding since this
results in fewer input channels.

Reviewers: godfrey.da.costa

Reviewed By: godfrey.da.costa

Differential Revision: https://phabricator.sourcevertex.net/D2500