Allow values of inChansPerGroup that don't divide the number of channels.
Summary:
If inChansPerGroup doesn't divide the number of channels we apply zero
padding to the inputs and the weights - take this into account when
estimating the transformation cost.

This fixes a performance regression in the weight update phase of ResNet-50
where, after expanding dimensions, there were some 1x1 convolutions where
the number of input channels was a multiple of 4 but not a multiple of 16.
Previously we set inChansPerGroup to 4 and under utilised the AMP unit but
in these cases it is better to zero pad.

Reviewers: godfrey.da.costa

Reviewed By: godfrey.da.costa

Differential Revision: https://phabricator.sourcevertex.net/D2495