Add a redundant constraint to speed up the planner.
Summary:
We can quickly estimate a lower bound on the number of cycles required
based on the maximum MACs per cycle per tile. The constraint solver
can use this additional constraint to prune the search space early,
speeding up the planning stage.

On my machine this reduces the amount of time it takes to build the
ResNet-50 graph from 84 seconds to 31 seconds.

Depends on D2504, D2505, D2506
Ref T1750

Reviewers: davel

Reviewed By: davel

Maniphest Tasks: T1750

Differential Revision: https://phabricator.sourcevertex.net/D2507