Rearrange constraint to reduce the chance hitting domain limits.
Summary:
Previously the maxMACs variable was defined to be the product of
maxMacsPerCyclePerTileVar, usedTiles and partialCalcCycles. This implicitly
constrained the product to be at most MAX_UINT - 1 (since the domain of
each variable is unsigned int and we don't allow variables to take the
value MAX_UINT). For large layers the product of these variables could
exceed this limit, in which case the planner would find no solution. We hit
this limit with a fully connected layer with --batch-size=128
--input-size=8192 --output-size=2048.

Rearrange the constraint to avoid the multiplication by
maxMacsPerCyclePerTileVar. Although this doesn't completely solve the
problem it increases the size of layer we can handle before it becomes
a problem.

Reviewers: davel

Reviewed By: davel

Differential Revision: https://phabricator.sourcevertex.net/D2531