Teach the planner to swap operands if profitable.
Summary:
Give the planner the freedom to swap operands. Swapping the operands has
the effect of changing which operand is loaded into the AMP unit which may
result in a more efficient plan.

The effect on ResNet training is as follows:

| Network   | Batch size | TFLOPS Before | TFLOPS After |
| ResNet-18 | 1          | 24.7          | 25.2         |
| ResNet-18 | 4          | 40.9          | 41.6         |
| ResNet-50 | 1          | 21.2          | 21.4         |

| Network   | Batch size | Memory Before | Memory After |
| ResNet-18 | 1          | 135.3         | 140.5        |
| ResNet-18 | 4          | 289.0         | 286.4        |
| ResNet-50 | 1          | 271.1         | 269.2        |

Note the convolution() function is still asymmetric - the cost model
applies a penalty to swapping the operands due to the rearrangement
so swapping the operands before calling convolution() may result in a
different plan.

Fixes T1752

Reviewers: godfrey.da.costa

Reviewed By: godfrey.da.costa

Maniphest Tasks: T1752

Differential Revision: https://phabricator.sourcevertex.net/D2525