Avoid allocations in the scheduler.
Summary:
Previously an unordered_set was used to check if a constraint was in
the worklist. This resulted in memory allocation / dellocations as entries
were added to / removed from the worklist. Replace this with
a std::vector<bool> indexed by the constraint number.

On ResNet-18 training this reduces the time until engine creation from
51 seconds to 45 seconds.

Reviewers: davel

Reviewed By: davel

Differential Revision: https://phabricator.sourcevertex.net/D2543