Remove swapConvolutionalWeightUpdateParams.
Summary:
Update calculateWeightDeltas() so it calls convolution()
with the activations on the left hand side, matching the way
Tensorflow would call it. To avoid a regression use the pass
information to improve the transform cost estimate, in
particular penalize plans that don't swap the operands if it is
the weight update pass.

Also teach the planner that both inputs to the convolution must
be rearranged if this is a weight update convolution, regardless
of whether they are swapped / expanded / padded. This has
gives a slight improvement on ResNet-50 training:

| Category | Before | After |
| Memory   | 269.2  | 268.4 |
| TFLOPS   | 21.35  | 21.38 |

There is no change on ResNet-18.

Fixes T1752

Reviewers: davel

Reviewed By: davel

Maniphest Tasks: T1752

Differential Revision: https://phabricator.sourcevertex.net/D2540