Fix handling of convolutions where the kernel is larger than the field.
Summary:
We define the behaviour of convolution() to be that the smaller operand is
slid over the larger operand. If the kernel is larger than the field then
if we pick a kernel index and iterate over the inputs multiplied by the
weight at that index then we must iterate over the partial sums in
reverse as we update them.

Note calculateWeightDeltas() is broken if the kernel size is larger than
the field size. I've added error checking to getWeightUpdateParams() to
detect this case and limited the tests to just running the forward pass.

Fixes T1830

Reviewers: godfrey.da.costa

Reviewed By: godfrey.da.costa

Maniphest Tasks: T1830

Differential Revision: https://phabricator.sourcevertex.net/D2590