Don't swap operands if joint planning is used.
Summary:
The logic to derive the fully connected backward / weight update plans
from the forward plan assumes the RHS operand is re-used - it doesn't handle
the case where operands are swapped and the LHS operand is re-used instead.
For now disallow swapping operands in the planner if joint planning is
enabled.

Fixes T1908.

Reviewers: godfrey.da.costa

Reviewed By: godfrey.da.costa

Maniphest Tasks: T1908

Differential Revision: https://phabricator.sourcevertex.net/D2745