Change addFlattenedRegions to produce regions in more sensible order
Summary:
Add flattened regions iterates though a multi-dimensional slice
building up the regions in the slice. The previous order of iteration
produced slices in an order that was awkward for later
processing. This update iterates in a different order producing
regions in the underlying order of the tensor.

No functionality change.

Reviewers: richard

Reviewed By: richard

Differential Revision: https://phabricator.sourcevertex.net/D2780