Fix getInputRange calculation
Summary:
Previously when calling getInputRange for a given kernel range it
iterated through the kernel range until it got to a kernel index that
gave a valid input range. However, this is not always the min/max
value of the input range when there is padding and kernel dilation.

Update the logic to try all elements of the kernel range but exit if a
definite min/max is found.

Test Plan:
This bug causes a failure in poplib tests when the new convolution
layout it used.

Reviewers: godfrey.da.costa

Reviewed By: godfrey.da.costa

Differential Revision: https://phabricator.sourcevertex.net/D2779