Add 1x1 convolution vertex for contiguous fields
Summary:
Vertex added by Dave L

The cost function is not called from the planner as the nx1 vertex will share the same code. The cost to zero out the partials is not included but a specialsed vertex which loads zeros for partials may be written without any significant additional cost except for calling the appropriate specialisation.

Reviewers: davel

Reviewed By: davel

Differential Revision: https://phabricator.sourcevertex.net/D2796