DynamicSlice from the sliciest dimension first
Summary: Rather than using the dimensions of the input tensor to decide the slice order, use the percentage reduction in tensor size.

Test Plan: All tests still pass. I added a very fragile test to check it does what I think it should which passes. If it breaks in future the test can be removed

Reviewers: Eddy

Reviewed By: Eddy

Differential Revision: https://phabricator.sourcevertex.net/D2841