Ensure integral output channels per group of reduced tensor are mapped to a tile
Summary:
The partial output channel groups of a multi-stage reduction of a convolution are ensured to be mapped on a tile so that any operation on a partial channel group is tile local. This is a workaround for a bug in xgraph which keeps tensors live if a group is split across more than a tile.

There is a small improvement in Memory and Cycles for both RESNET50 and RESNET18

RESNET-50 / No BatchNorm / Batch Size=1
       Mem             Cyc
Old    279088956       1640317
New    275831064       1617432

RESNET-50 / No BatchNorm / Batch Size=2
Old    407926700       2434172
New    405975816       2416510

RESNET-50 / BatchNorm / Batch Size=1
Old    317412876       2008674
New    309405344       1919819

RESNET-50 / BatchNorm / Batch Size=2
Old    465497824       2851624
New    459234648       2779473

Fixes Maniphest T1922

Reviewers: richard

Reviewed By: richard

Differential Revision: https://phabricator.sourcevertex.net/D2842