Increase float test tolerances.
Summary:
On the CPU model there is no difference in precision between float and half
(since half is just an alias for float by default) and therefore there is
no reason to use a difference tolerance. Other changes to poplibs I've been
working on perturb the BatchNormConv_Batch4_DataFloat_PartialsFloat test
such that the float tolerances are too tight - this change fixes that.

Reviewers: godfrey.da.costa

Reviewed By: godfrey.da.costa

Differential Revision: https://phabricator.sourcevertex.net/D2856