Update activation dimension ordering to NCHW.
Summary:
Update the convolution, batch norm and pooling functions to expect
tensor in the order NCHW. This change just changes the API - inside these
functions parameters are dimshuffled back to the old format before any
operations are performed. The order of dimensions of the weight tensor
are unchanged.

Ref T1679

Reviewers: davel

Reviewed By: davel

Maniphest Tasks: T1679

Differential Revision: https://phabricator.sourcevertex.net/D2816