Fix handling of non co-prime input dilation and stride.
Summary:
Suppose the input dilation and the stride share a common factor.
If the lower input padding is not a multiple of the common factor the
output of the convolution is zero. We already handle this case correctly
as getOutputRange() returns an empty range and no entries are added to
the worklist. This just leaves cases where the lower input padding is a
multiple of the common factor. In these cases the convolution is
equivalent to a convolution where the dilation, padding and stride all
divided by this common factor. Divide the vertex in and out strides by
their common factor to ensure we get the correct results.

Fixes T2015

Reviewers: davel

Reviewed By: davel

Maniphest Tasks: T2015

Differential Revision: https://phabricator.sourcevertex.net/D3034