Fix calculation of the amount of input window padding required.
Summary:
If there is striding the first element of the input window of the
first tile may not be the first non padding element in the field in
which case the amount of padding we need to the input window could be
greater than the amount of padding specified for the layer. Replace the
previous logic with code that computes a simple conservative
upper bound on the amount of padding required.

Only apply padding to the input window if the convUnitWeightHeight is
greater than 1 - there is no reason to pad otherwise.

Reviewers: davel

Reviewed By: davel

Differential Revision: https://phabricator.sourcevertex.net/D3040