Extend popconv to support flipping the activations and the weights.
Summary:
We can now swap operands by adding padding to the weights and flipping
both the activations and the weights. This lets us drop support for kernels
larger than the input, simplifying the code in a number of places. Any
tests that used a large kernel have been changed to an equivalent
convolution that flips the activations and weights instead.

Fixes T1831

Reviewers: godfrey.da.costa

Reviewed By: godfrey.da.costa

Maniphest Tasks: T1831

Differential Revision: https://phabricator.sourcevertex.net/D3119