Update transform to increase the number of output channels.
Summary:
The transform to increase the number of output channels previously checked
if the input field had size 1 and the kernel had size greater than 1. As a
result the transform is never applied now that poplibs has been changed
such that the kernel size is always less than or equal to the input size.

Previously the transform typically followed an expansion a dimension of
a convolution where the kernel was larger than input field. We can get
the same result as the combined operation by swapping the operands,
expanding the field, flattening the dimension into the batch axis and
then swapping the operands back.

When the transform is implement this way we have the ability to flatten
any spatial dimension of any convolution into the output channel dimension
(previously we could only do this if the kernel size was larger than the
input size). To avoid increasing in the planner's search space we
constrain so it doesn't consider expanding the dimension if the
kernel is larger than the input and it doesn't consider flattening the
dimension into the output channel dimension if the input is larger than
the kernel.

Reviewers: davel

Reviewed By: davel

Differential Revision: https://phabricator.sourcevertex.net/D3189