Move tile partitioning information into a separate structure.
Summary:
As part of this change introduce fields for input / output channel
grain size. Previously the grain size was assumed to match the channel
grouping but conceptually these are different things - one determines how
the problem is split across tiles and the other controls the layout of
data on a tile. Currently the planner sets the grain size to the channel
grouping and so there is no functionality change.

Reviewers: godfrey.da.costa

Reviewed By: godfrey.da.costa

Differential Revision: https://phabricator.sourcevertex.net/D3249