Change cycle estimator function to not rely on host half type
Summary:
Cycle estimator functions were templated on host types to reflect
codelete templates in device code. The host 'half' type is going to be
removed from poplar so this commit changes the functions to be
parameterized by device types instead (i.e. the poplar::Type type).

This changes template parameters to function parameters. The helper
macros/functions in cyclesTables.hpp are changed to cope with this.

One issue is that the parameters are globals whose initialization order
is undefined w.r.t other globals so the global cycle table structures
are changes to functions that produce the cycle table to ensure the
table is constructed after the poplar::Type objects.

Reviewers: markp

Reviewed By: markp

Maniphest Tasks: T2167

Differential Revision: https://phabricator.sourcevertex.net/D3296