Remove support for a variable number of product constraint operands.
Summary:
Limit the product constraint to the product of two variables.
Implement constraints on products of more than two variables using
temporary variables and multiple product constraints. This simplifies the
product constraint avoiding the need for satMul function.

When running single_conv_layer targeting 2 IPUS for a layer in the middle
of ResNet-50 this reduces the time before engine creation from 12.2
seconds to 10.6 seconds.

Depends on D3245

Reviewers: davel

Reviewed By: davel

Differential Revision: https://phabricator.sourcevertex.net/D3301