Express exchange cost using simple constraints.
Summary:
Express the exchange cost using a combination of the product, sum and
ceildiv constraints on intermediate variables instead of an opaque call.
This exposes more information that the constraint solver can use to rule
out parts of the search space as it is tries to solve the optimization
problem.

This provides a modest reduction in the compile time when there is one
IPU, for example on my machine the total compile time for ResNet-18
training (1 IPU) is reduced from 109 seconds to 103 seconds. Greater
reductions in compile time are seen for multi-IPU targets. For example the
time it takes to run single_conv_layer for a layer in the middle of
ResNet-50 when targetting a 2 IPU device is reduced from 149 seconds to
44 seconds.

Depends on D3251

Reviewers: davel

Reviewed By: davel

Differential Revision: https://phabricator.sourcevertex.net/D3276