For each tile transform ConvParams to describe the sub-convolution.
Summary:
Transforming ConvParams should simplify vertex creation code as it lets us
query information about the convolution performed on the tile without
having to separately maintain information about the slice the tile is
operating on. The vertex creation code is unchanged for now - I intend to
update it in a separate commit.

Ref T1737

Reviewers: davel

Reviewed By: davel

Maniphest Tasks: T1737

Differential Revision: https://phabricator.sourcevertex.net/D3507