Make channel padding explicit in the plan.
Summary:
Previously the number of channels was implicitly padded to a multiple
of the channel grouping. This change makes the amount of padding added
explicit in the plan. In the future when there are transforms at different
levels of the hierarchy we can use the padding fields to specify which
level at add padding at.

Reviewers: godfrey.da.costa

Reviewed By: godfrey.da.costa

Differential Revision: https://phabricator.sourcevertex.net/D3557